requirements.txt
openai
langchain
pinecone-client
unstructured
pdf2image
pytesseract
tiktoken
streamlit
evaluate
pythainlp
transformers
fairseq
sacremoses
sentencepiece